refactor: 영상도네 패키지 구조 변경
