+++ src/main/java/com/chzzkGamble/videodonation/youtube/YoutubeClient.java
@@ -1,8 +1,8 @@
-package com.chzzkGamble.videodonation;
+package com.chzzkGamble.videodonation.youtube;
 
 import com.chzzkGamble.exception.ChzzkException;
 import com.chzzkGamble.exception.ChzzkExceptionCode;
-import com.chzzkGamble.videodonation.YouTubeApiResponse.Item;
+import com.chzzkGamble.videodonation.youtube.YouTubeApiResponse.Item;
 import lombok.AllArgsConstructor;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.stereotype.Component;
